[Frontend] Refactor ServingTokenization entrypoint. (#46022)
