+++ vllm/entrypoints/openai/engine/serving.py
@@ -5,25 +5,19 @@ import time
 from collections.abc import Awaitable, Mapping
 from dataclasses import dataclass, field
 from http import HTTPStatus
-from typing import Any, ClassVar, Generic, Protocol, TypeAlias, TypeVar
+from typing import ClassVar, Generic, TypeVar
 
 from fastapi import Request
 from pydantic import ConfigDict
 from starlette.datastructures import Headers
 
-import vllm.envs as envs
-from vllm.config import ModelConfig
 from vllm.engine.protocol import EngineClient
-from vllm.entrypoints.chat_utils import ChatTemplateContentFormatOption
 from vllm.entrypoints.generate.beam_search.online import BeamSearchOnlineMixin
 from vllm.entrypoints.openai.chat_completion.protocol import (
-    BatchChatCompletionRequest,
     ChatCompletionRequest,
-    ChatCompletionResponse,
 )
 from vllm.entrypoints.openai.completion.protocol import (
     CompletionRequest,
-    CompletionResponse,
 )
 from vllm.entrypoints.openai.engine.protocol import (
     ErrorResponse,
@@ -31,80 +25,21 @@ from vllm.entrypoints.openai.engine.protocol import (
 )
 from vllm.entrypoints.openai.models.serving import OpenAIServingModels
 from vllm.entrypoints.openai.responses.protocol import ResponsesRequest
-from vllm.entrypoints.serve.disagg.protocol import GenerateRequest, GenerateResponse
-from vllm.entrypoints.serve.tokenize.protocol import (
-    DetokenizeRequest,
-    TokenizeChatRequest,
-    TokenizeCompletionRequest,
-    TokenizeResponse,
-)
-from vllm.entrypoints.serve.utils.error_response import create_error_response
+from vllm.entrypoints.serve.engine.serving import BaseServing
+from vllm.entrypoints.serve.engine.typing import AnyRequest
 from vllm.entrypoints.serve.utils.request_logger import RequestLogger
-from vllm.entrypoints.speech_to_text.transcription.protocol import (
-    TranscriptionRequest,
-    TranscriptionResponse,
-)
-from vllm.entrypoints.speech_to_text.translation.protocol import TranslationRequest
-from vllm.inputs import EngineInput, PromptType
+from vllm.inputs import EngineInput
 from vllm.logger import init_logger
 from vllm.logprobs import Logprob, PromptLogprobs
 from vllm.lora.request import LoRARequest
-from vllm.renderers import ChatParams, TokenizeParams
-from vllm.renderers.inputs.preprocess import (
-    extract_prompt_components,
-    extract_prompt_len,
-)
-from vllm.sampling_params import BeamSearchParams, SamplingParams
 from vllm.tokenizers import TokenizerLike
 from vllm.tracing import (
     contains_trace_headers,
     extract_trace_headers,
     log_tracing_disabled_warning,
 )
-from vllm.utils import random_uuid
 
 logger = init_logger(__name__)
-
-
-class RendererRequest(Protocol):
-    def build_tok_params(self, model_config: ModelConfig) -> TokenizeParams:
-        raise NotImplementedError
-
-
-class RendererChatRequest(RendererRequest, Protocol):
-    def build_chat_params(
-        self,
-        default_template: str | None,
-        default_template_content_format: ChatTemplateContentFormatOption,
-    ) -> ChatParams:
-        raise NotImplementedError
-
-
-CompletionLikeRequest: TypeAlias = (
-    CompletionRequest | TokenizeCompletionRequest | DetokenizeRequest
-)
-
-ChatLikeRequest: TypeAlias = (
-    ChatCompletionRequest | BatchChatCompletionRequest | TokenizeChatRequest
-)
-
-SpeechToTextRequest: TypeAlias = TranscriptionRequest | TranslationRequest
-
-AnyRequest: TypeAlias = (
-    CompletionLikeRequest
-    | ChatLikeRequest
-    | SpeechToTextRequest
-    | ResponsesRequest
-    | GenerateRequest
-)
-
-AnyResponse: TypeAlias = (
-    CompletionResponse
-    | ChatCompletionResponse
-    | TranscriptionResponse
-    | TokenizeResponse
-    | GenerateResponse
-)
 
 RequestT = TypeVar("RequestT", bound=AnyRequest)
 _T = TypeVar("_T")
@@ -122,7 +57,7 @@ class ServeContext(Generic[RequestT]):
     model_config = ConfigDict(arbitrary_types_allowed=True)
 
 
-class OpenAIServing(BeamSearchOnlineMixin):
+class OpenAIServing(BaseServing, BeamSearchOnlineMixin):
     request_id_prefix: ClassVar[str] = """
     A short string prepended to every request’s ID.
     """
@@ -135,15 +70,14 @@ class OpenAIServing(BeamSearchOnlineMixin):
         request_logger: RequestLogger | None,
         return_tokens_as_token_ids: bool = False,
     ):
-        super().__init__()
+        super().__init__(
+            models=models,
+            model_config=engine_client.model_config,
+            request_logger=request_logger,
+        )
 
         self.engine_client = engine_client
-        self.models = models
-
-        self.request_logger = request_logger
         self.return_tokens_as_token_ids = return_tokens_as_token_ids
-
-        self.model_config = engine_client.model_config
         self.renderer = engine_client.renderer
         self.input_processor = engine_client.input_processor
         vllm_config = getattr(engine_client, "vllm_config", None)
@@ -162,15 +96,6 @@ class OpenAIServing(BeamSearchOnlineMixin):
         except Exception:
             # Never fail server startup over the fingerprint.
             self.system_fingerprint = None
-
-    @staticmethod
-    def create_error_response(
-        message: str | Exception,
-        err_type: str = "BadRequestError",
-        status_code: HTTPStatus = HTTPStatus.BAD_REQUEST,
-        param: str | None = None,
-    ) -> ErrorResponse:
-        return create_error_response(message, err_type, status_code, param)
 
     def create_streaming_error_response(
         self,
@@ -208,167 +133,6 @@ class OpenAIServing(BeamSearchOnlineMixin):
             status_code=e.status_code,
         )
 
-    async def _check_model(
-        self,
-        request: AnyRequest,
-    ) -> ErrorResponse | None:
-        error_response = None
-
-        if self._is_model_supported(request.model):
-            return None
-        if request.model in self.models.lora_requests:
-            return None
-        if (
-            envs.VLLM_ALLOW_RUNTIME_LORA_UPDATING
-            and request.model
-            and (load_result := await self.models.resolve_lora(request.model))
-        ):
-            if isinstance(load_result, LoRARequest):
-                return None
-            if (
-                isinstance(load_result, ErrorResponse)
-                and load_result.error.code == HTTPStatus.BAD_REQUEST.value
-            ):
-                error_response = load_result
-
-        return error_response or self.create_error_response(
-            message=f"The model `{request.model}` does not exist.",
-            err_type="NotFoundError",
-            status_code=HTTPStatus.NOT_FOUND,
-            param="model",
-        )
-
-    def _get_active_default_mm_loras(self, request: AnyRequest) -> LoRARequest | None:
-        """Determine if there are any active default multimodal loras."""
-        # TODO: Currently this is only enabled for chat completions
-        # to be better aligned with only being enabled for .generate
-        # when run offline. It would be nice to support additional
-        # tasks types in the future.
-        message_types = self._get_message_types(request)
-        default_mm_loras = set()
-
-        for lora in self.models.lora_requests.values():
-            # Best effort match for default multimodal lora adapters;
-            # There is probably a better way to do this, but currently
-            # this matches against the set of 'types' in any content lists
-            # up until '_', e.g., to match audio_url -> audio
-            if lora.lora_name in message_types:
-                default_mm_loras.add(lora)
-
-        # Currently only support default modality specific loras if
-        # we have exactly one lora matched on the request.
-        if len(default_mm_loras) == 1:
-            return default_mm_loras.pop()
-        return None
-
-    def _maybe_get_adapters(
-        self,
-        request: AnyRequest,
-        supports_default_mm_loras: bool = False,
-    ) -> LoRARequest | None:
-        if request.model in self.models.lora_requests:
-            return self.models.lora_requests[request.model]
-
-        # Currently only support default modality specific loras
-        # if we have exactly one lora matched on the request.
-        if supports_default_mm_loras:
-            default_mm_lora = self._get_active_default_mm_loras(request)
-            if default_mm_lora is not None:
-                return default_mm_lora
-
-        if self._is_model_supported(request.model):
-            return None
-
-        # if _check_model has been called earlier, this will be unreachable
-        raise ValueError(f"The model `{request.model}` does not exist.")
-
-    def _get_message_types(self, request: AnyRequest) -> set[str]:
-        """Retrieve the set of types from message content dicts up
-        until `_`; we use this to match potential multimodal data
-        with default per modality loras.
-        """
-        message_types: set[str] = set()
-
-        if not hasattr(request, "messages"):
-            return message_types
-
-        messages = request.messages
-        if messages is None or isinstance(messages, (str, bytes)):
-            return message_types
-
-        for message in messages:
-            if (
-                isinstance(message, dict)
-                and "content" in message
-                and isinstance(message["content"], list)
-            ):
-                for content_dict in message["content"]:
-                    if "type" in content_dict:
-                        message_types.add(content_dict["type"].split("_")[0])
-        return message_types
-
-    def _validate_chat_template(
-        self,
-        request_chat_template: str | None,
-        chat_template_kwargs: dict[str, Any] | None,
-        trust_request_chat_template: bool,
-    ) -> ErrorResponse | None:
-        if not trust_request_chat_template and (
-            request_chat_template is not None
-            or (
-                chat_template_kwargs
-                and chat_template_kwargs.get("chat_template") is not None
-            )
-        ):
-            return self.create_error_response(
-                "Chat template is passed with request, but "
-                "--trust-request-chat-template is not set. "
-                "Refused request with untrusted chat template."
-            )
-        return None
-
-    @staticmethod
-    def _prepare_extra_chat_template_kwargs(
-        request_chat_template_kwargs: dict[str, Any] | None = None,
-        default_chat_template_kwargs: dict[str, Any] | None = None,
-    ) -> dict[str, Any]:
-        """Helper to merge server-default and request-specific chat template kwargs."""
-        request_chat_template_kwargs = request_chat_template_kwargs or {}
-        if default_chat_template_kwargs is None:
-            return request_chat_template_kwargs
-        # Apply server defaults first, then request kwargs override.
-        return default_chat_template_kwargs | request_chat_template_kwargs
-
-    def _extract_prompt_components(self, prompt: PromptType | EngineInput):
-        return extract_prompt_components(self.model_config, prompt)
-
-    def _extract_prompt_text(self, prompt: PromptType | EngineInput):
-        return self._extract_prompt_components(prompt).text
-
-    def _extract_prompt_len(self, prompt: EngineInput):
-        return extract_prompt_len(self.model_config, prompt)
-
-    def _log_inputs(
-        self,
-        request_id: str,
-        inputs: PromptType | EngineInput,
-        params: SamplingParams | BeamSearchParams | None,
-        lora_request: LoRARequest | None,
-    ) -> None:
-        if self.request_logger is None:
-            return
-
-        components = self._extract_prompt_components(inputs)
-
-        self.request_logger.log_inputs(
-            request_id,
-            components.text,
-            components.token_ids,
-            components.embeds,
-            params=params,
-            lora_request=lora_request,
-        )
-
     async def _get_trace_headers(
         self,
         headers: Headers,
@@ -382,18 +146,6 @@ class OpenAIServing(BeamSearchOnlineMixin):
             log_tracing_disabled_warning()
 
         return None
-
-    @staticmethod
-    def _base_request_id(
-        raw_request: Request | None, default: str | None = None
-    ) -> str | None:
-        """Pulls the request id to use from a header, if provided"""
-        if raw_request is not None and (
-            (req_id := raw_request.headers.get("X-Request-Id")) is not None
-        ):
-            return req_id
-
-        return random_uuid() if default is None else default
 
     @staticmethod
     def _get_data_parallel_rank(raw_request: Request | None) -> int | None:
@@ -463,13 +215,6 @@ class OpenAIServing(BeamSearchOnlineMixin):
             )
 
         return tokenizer.decode([token_id])
-
-    def _is_model_supported(self, model_name: str | None) -> bool:
-        if not model_name:
-            return True
-        if envs.VLLM_SKIP_MODEL_NAME_VALIDATION:
-            return True
-        return self.models.is_base_model(model_name)
 
 
 def format_token_id_placeholder(token_id: int) -> str:
