+++ vllm/entrypoints/serve/disagg/api_router.py
@@ -20,7 +20,7 @@ from vllm.entrypoints.serve.disagg.protocol import (
 from vllm.entrypoints.serve.disagg.serving import (
     ServingTokens,
 )
-from vllm.entrypoints.serve.tokenize.serving import OpenAIServingTokenization
+from vllm.entrypoints.serve.tokenize.serving import ServingTokenization
 from vllm.entrypoints.serve.utils.api_utils import (
     load_aware_call,
     validate_json_request,
@@ -31,8 +31,8 @@ from vllm.logger import init_logger
 logger = init_logger(__name__)
 
 
-def tokenization(request: Request) -> OpenAIServingTokenization:
-    return request.app.state.openai_serving_tokenization
+def tokenization(request: Request) -> ServingTokenization:
+    return request.app.state.serving_tokenization
 
 
 def generate_tokens(request: Request) -> ServingTokens | None:
